Refactor: character 관련 repository 클래스 패키지 삭제
