+++ src/main/java/com/fours/tolevelup/service/MissionService.java
@@ -9,8 +9,8 @@ import com.fours.tolevelup.model.entity.Mission;
 import com.fours.tolevelup.model.entity.MissionLog;
 import com.fours.tolevelup.model.entity.User;
 import com.fours.tolevelup.model.entity.UserCharacter;
-import com.fours.tolevelup.repository.character.CharacterRepository;
-import com.fours.tolevelup.repository.character.UserCharacterRepository;
+import com.fours.tolevelup.repository.CharacterRepository;
+import com.fours.tolevelup.repository.UserCharacterRepository;
 import com.fours.tolevelup.repository.MissionRepository;
 import com.fours.tolevelup.repository.MissionLogRepository;
 import com.fours.tolevelup.repository.ThemeExpRepository;
