+++ src/main/java/com/fours/tolevelup/service/UserService.java
@@ -17,8 +17,8 @@ import com.fours.tolevelup.repository.AlarmRepository;
 import com.fours.tolevelup.repository.CommentRepository;
 import com.fours.tolevelup.repository.FollowRepository;
 import com.fours.tolevelup.repository.LikeRepository;
-import com.fours.tolevelup.repository.character.CharacterRepository;
-import com.fours.tolevelup.repository.character.UserCharacterRepository;
+import com.fours.tolevelup.repository.CharacterRepository;
+import com.fours.tolevelup.repository.UserCharacterRepository;
 import com.fours.tolevelup.repository.MissionLogRepository;
 import com.fours.tolevelup.repository.ThemeRepository;
 import com.fours.tolevelup.repository.ThemeExpRepository;
