+++ src/main/java/com/fours/tolevelup/service/character/CharacterService.java
@@ -3,8 +3,8 @@ package com.fours.tolevelup.service.character;
 import com.fours.tolevelup.controller.request.UserCharacterRequest;
 import com.fours.tolevelup.model.entity.Character;
 import com.fours.tolevelup.model.entity.UserCharacter;
-import com.fours.tolevelup.repository.character.CharacterRepository;
-import com.fours.tolevelup.repository.character.UserCharacterRepository;
+import com.fours.tolevelup.repository.CharacterRepository;
+import com.fours.tolevelup.repository.UserCharacterRepository;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
